+++ src/share/theom/config/i3/osd/volume-change.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
-enable_osd=$(theom-config features.osd | tr -d "[:space:]")
+enable_osd=$(theom-config osd.osd | tr -d "[:space:]")
 
 if [[ "$enable_osd" == "true" ]]; then
-    ~/.config/.theom/osd-snippets/volume.sh
+    ~/.config/theom/.osd-snippets/volume.sh
 fi