+++ src/share/theom/config/i3/other/first-time-welcome.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
-TSCL_PATH="$HOME/.config/.THEOM_SYSTEM/&info&.tscd"
-TSC="tscdctl"
+TSCL_PATH="$HOME/.local/state/theom/meta.tscd"
+TSC="/usr/share/theom/scripts/tscdctl"
 TSCD_value=$($TSC --get current_run --out "$TSCL_PATH")
 
 if [ "$TSCD_value" = "True" ]; then